Shall I change my mind, shall I not change my mind

+++ src/array.h
@@ -157,9 +157,9 @@ typedef struct {
 	int last_used_ndx; /* round robin */
 
 	/* brownout */
-	int lastLastTheta; /* 0 to 10000 */
-	int lastTheta; /* 0 to 10000 */
-	int weight; /* 0 to 10000 */
+	float lastLastTheta;
+	float lastTheta;
+	int weight;
 } data_fastcgi;
 
 data_fastcgi *data_fastcgi_init(void);
